Splitted the OSC Controller class
